fastapi/app/requirements.txt
fastapi
uvicorn
pydantic-settings
sqlmodel
psycopg2-binary
alembic
passlib[bcrypt]
python-multipart
python-jose[cryptography]